emotion-detector/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
ibm-watson>=8.0.0
python-dotenv>=1.0.1
requests>=2.32.3
pytest>=8.2.0
flake8>=7.1.0
pylint>=3.2.0
